summary:
Added a home-button in the form of a logo, added a responsive layout, cleaned up some of the javascript code (noticed that I had written a few redundant lines), and cleaned up a lot of the CSS. only an e-mail form remains now and whatever design-choices we make together throughout the week. the proportions of the layout based on screen-resolution can be easily altered at this stage.